Break out all slice division functions.
